+++ commerce-domain/src/main/java/co/hooghly/commerce/repository/CustomerAttributeRepository.java
@@ -14,14 +14,14 @@ public interface CustomerAttributeRepository extends JpaRepository<CustomerAttri
 	CustomerAttribute findOne(Long id);
 	
 	@Query("select a from CustomerAttribute a join fetch a.customer ac left join fetch a.customerOption aco join fetch aco.merchantStore acom left join fetch a.customerOptionValue acov left join fetch aco.descriptions acod left join fetch acov.descriptions acovd where acom.id = ?1 and ac.id = ?2 and aco.id = ?3")
-	CustomerAttribute findByOptionId(Integer merchantId,Long customerId,Long id);
+	CustomerAttribute findByOptionId(Long merchantId,Long customerId,Long id);
 	
 	@Query("select a from CustomerAttribute a join fetch a.customer ac left join fetch a.customerOption aco join fetch aco.merchantStore acom left join fetch a.customerOptionValue acov left join fetch aco.descriptions acod left join fetch acov.descriptions acovd where acom.id = ?1 and aco.id = ?2")
-	List<CustomerAttribute> findByOptionId(Integer merchantId,Long id);
+	List<CustomerAttribute> findByOptionId(Long merchantId,Long id);
 
 	@Query("select distinct a from CustomerAttribute a join fetch a.customer ac left join fetch a.customerOption aco join fetch aco.merchantStore acom left join fetch a.customerOptionValue acov left join fetch aco.descriptions acod left join fetch acov.descriptions acovd where acom.id = ?1 and ac.id = ?2")
-	List<CustomerAttribute> findByCustomerId(Integer merchantId,Long customerId);
+	List<CustomerAttribute> findByCustomerId(Long merchantId,Long customerId);
 	
 	@Query("select a from CustomerAttribute a join fetch a.customer ac left join fetch a.customerOption aco join fetch aco.merchantStore acom left join fetch a.customerOptionValue acov left join fetch aco.descriptions acod left join fetch acov.descriptions acovd where acom.id = ?1 and acov.id = ?2")
-	List<CustomerAttribute> findByOptionValueId(Integer merchantId,Long Id);
+	List<CustomerAttribute> findByOptionValueId(Long merchantId,Long Id);
 }
